requirements.txt
alembic==1.0.11
aniso8601==8.0.0
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.11.28
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-JSGlue==0.3.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SocketIO==4.2.0
Flask-SQLAlchemy==2.4.0
Flask-Testing==0.7.1
Flask-WTF==0.14.2
idna==2.8
importlib-metadata==1.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.2.0
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.1.0
ndg-httpsclient==0.5.1
passlib==1.7.1
pbr==5.4.4
pgpubsub==0.0.5
psycopg2==2.7.7
pusher==2.1.4
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyrsistent==0.15.7
python-dateutil==2.8.0
python-editor==1.0.4
python-engineio==3.9.2
python-socketio==4.3.0
pytz==2019.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.6
sqlalchemy-json==0.2.3
SQLAlchemy-Utils==0.34.1
stevedore==1.31.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
Werkzeug==0.15.5
WTForms==2.2.1
zipp==2.0.0
